+++ src/globalScript/lighthouse/core.tsx
@@ -1,0 +1,33 @@
+import { createContext } from '@wordpress/element'
+import { BehaviorSubject } from 'rxjs'
+import { some } from './option'
+import type { Action, Root } from './types'
+import { reducer } from './reducer'
+
+export const FILTER_PRIVATE = '__'
+export const BlockIdContext = createContext('')
+
+export const blockStoreSubject = new BehaviorSubject<Root>({
+	__root: {
+		blockName: 'core',
+		blockType: 'inseri-core/root',
+		clientId: '',
+		state: 'ready',
+		atoms: {
+			'detailed-data-flow': { description: 'detailed data-flow', content: some({ contentType: 'application/json', value: [] }) },
+			'data-flow': { description: 'data-flow', content: some({ contentType: 'application/json', value: [] }) },
+			blocks: { description: 'blocks', content: some({ contentType: 'application/json', value: [] }) },
+			'is-hidden': { description: 'are blocks hidden', content: some({ contentType: 'application/json', value: false }) },
+		},
+	},
+})
+
+export function onNext(action: Action) {
+	const reducedBase = reducer(blockStoreSubject.getValue(), action)
+	blockStoreSubject.next(reducedBase)
+}
+
+if (process.env.NODE_ENV !== 'production') {
+	// eslint-disable-next-line no-console
+	blockStoreSubject.subscribe((root) => console.log('#lighthouse:', root))
+}
