+++ src/globalScript/lighthouse/index.tsx
@@ -1,0 +1,8 @@
+import { InseriRoot } from './InseriRoot'
+import { useDiscover, DiscoveredItem } from './useDiscover'
+import { usePublish, Actions } from './usePublish'
+import { useWatch } from './useWatch'
+import type { Nucleus } from './types'
+
+export { InseriRoot, useDiscover, usePublish, useWatch }
+export type { Actions, Nucleus, DiscoveredItem }
