+++ src/globalScript/lighthouse/option.ts
