+++ src/globalScript/lighthouse/reducer.ts
