+++ src/globalScript/lighthouse/types.ts
